fix: root 프로젝트 명 및 패키지 변경
