+++ module-core/module-client/build.gradle.kts
@@ -5,26 +5,10 @@ plugins {
     id("io.spring.dependency-management")
 }
 
-group = "com.almumol"
-version = "0.0.1-SNAPSHOT"
-
-repositories {
-    mavenCentral()
-}
-
 dependencies {
     implementation(project(":module-core:module-domain"))
 
     implementation("org.springframework.boot:spring-boot-starter-web")
-
-    testImplementation(kotlin("test"))
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-kotlin {
-    jvmToolchain(21)
 }
 
 tasks.getByName<org.springframework.boot.gradle.tasks.bundling.BootJar>("bootJar") {
