+++ module-core/module-client/src/main/kotlin/almumol/ossori/core/client/dto/response/GithubRepositoriesResponse.kt
@@ -1,0 +1,13 @@
+package almumol.ossori.core.client.dto.response
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class GithubRepositoriesResponse(
+
+    @JsonProperty("total_count")
+    val totalCount: Long,
+
+    val items: List<GithubRepositoryResponse>
+
+) {
+}
