+++ module-core/module-client/src/main/kotlin/almumol/ossori/core/client/github/GithubClientProperties.kt
@@ -1,0 +1,14 @@
+package almumol.ossori.core.client.github
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "security.github")
+data class GithubClientProperties(
+
+    val repositoryBaseUrl: String,
+
+    val searchRepositoryBaseUrl: String,
+
+    val token: String
+) {
+}
