+++ module-core/module-client/src/main/kotlin/almumol/ossori/core/client/global/exception/GithubResponseException.kt
@@ -1,0 +1,7 @@
+package almumol.ossori.core.client.global.exception
+
+import org.springframework.http.HttpStatusCode
+
+class GithubResponseException(statusCode: HttpStatusCode, message: String) : RuntimeException(
+    "Status code: ${statusCode}, Description: $message"
+)
