--- module-core/module-client/src/main/kotlin/client/github/GithubClientProperties.kt
@@ -1,14 +1,0 @@
-package client.github
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties(prefix = "security.github")
-data class GithubClientProperties(
-
-    val repositoryBaseUrl: String,
-
-    val searchRepositoryBaseUrl: String,
-
-    val token: String
-) {
-}
