--- module-core/module-client/src/main/kotlin/config/RestClientConfig.kt
@@ -1,30 +1,0 @@
-package config
-
-import client.github.GithubResponseErrorHandler
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.http.client.ClientHttpRequestFactory
-import org.springframework.http.client.SimpleClientHttpRequestFactory
-import org.springframework.web.client.RestClient
-import java.time.Duration
-
-@Configuration
-class RestClientConfig {
-
-    @Bean
-    fun githubRestClient(): RestClient {
-        val requestFactory = simpleRequestFactory()
-        return RestClient.builder()
-            .requestFactory(requestFactory)
-            .defaultStatusHandler(GithubResponseErrorHandler())
-            .build()
-    }
-
-    private fun simpleRequestFactory(): ClientHttpRequestFactory {
-        val requestFactory = SimpleClientHttpRequestFactory()
-        requestFactory.setConnectTimeout(Duration.ofSeconds(2))
-        requestFactory.setReadTimeout(Duration.ofSeconds(10))
-
-        return requestFactory
-    }
-}
