--- module-core/module-client/src/main/kotlin/dto/response/GithubRepositoriesResponse.kt
@@ -1,13 +1,0 @@
-package dto.response
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class GithubRepositoriesResponse(
-
-    @JsonProperty("total_count")
-    val totalCount: Long,
-
-    val items: List<GithubRepositoryResponse>
-
-) {
-}
