--- module-core/module-client/src/main/kotlin/global/exception/GithubResponseException.kt
@@ -1,7 +1,0 @@
-package global.exception
-
-import org.springframework.http.HttpStatusCode
-
-class GithubResponseException(statusCode: HttpStatusCode, message: String) : RuntimeException(
-    "Status code: ${statusCode}, Description: $message"
-)
