+++ module-core/module-client/src/test/kotlin/client/github/GithubClientTest.kt
@@ -1,5 +1,6 @@
 package client.github
 
+import almumol.ossori.core.client.github.GithubClient
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.extension.ExtendWith
@@ -12,7 +13,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension
 import kotlin.test.Test
 
 @TestConfiguration
-@ComponentScan(basePackages = ["client", "config"])
+@ComponentScan(basePackages = ["client"])
 class TestConfig() {
 }
