+++ module-core/module-domain/src/main/kotlin/almumol/ossori/core/domain/project/global/exception/BadRequestException.kt
@@ -1,0 +1,3 @@
+package almumol.ossori.core.domain.project.global.exception
+
+data class BadRequestException(override val message: String) : RuntimeException()
