+++ module-core/module-domain/src/main/kotlin/almumol/ossori/core/domain/project/global/exception/ExceptionResponse.kt
@@ -1,0 +1,5 @@
+package almumol.ossori.core.domain.project.global.exception
+
+data class ExceptionResponse(
+        val message: String
+)
