+++ module-core/module-domain/src/main/kotlin/almumol/ossori/core/domain/project/repository/ProjectRepository.kt
@@ -1,0 +1,9 @@
+package almumol.ossori.core.domain.project.repository
+
+import almumol.ossori.core.domain.project.domain.Project
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ProjectRepository : JpaRepository<Project, Long> {
+}
