+++ settings.gradle.kts
@@ -1,9 +1,8 @@
-rootProject.name = "OSSori"
+rootProject.name = "ossori"
 include(
     ":module-api",
     ":module-batch",
     ":module-core:module-domain",
     ":module-core:module-client",
-    ":module-core:module-infra",
-    "module-frontend"
+    ":module-core:module-infra"
 )
